requirements.txt

Django
djangorestframework
psycopg2
drf-spectacular
Pillow==11.1.0